requirements.txt
streamlit
pandasai
python-dotenv
bs4
langchain
seaborn